# app/Repositories/Api/V1/Admin/Reset/AdminUserResetRepository.php
<?php

namespace App\Repositories\Api\V1\Admin\Reset;

use App\Repositories\BaseRepository;
use App\Models\AdminUser;
use App\Traits\BaseTrait;
use Carbon\Carbon;
use Illuminate\Http\Request;
use Illuminate\Http\JsonResponse;
use Hash;
use Mail;
//vpx_imports
class AdminUserResetRepository extends BaseRepository implements IAdminUserResetRepository
{
    use BaseTrait;
    public function __construct()
    {

        $this->lang = 'admin.reset';
    }

    /**
     * View login page
     *
     * @param Request $request
     * @return View
     */
    public function index(Request $request): JsonResponse
    {
        $data['lang'] = $this->lang;
        return response()->json(['success' => true, 'data' => $data]);
    }

    public function sendCode(Request $request): JsonResponse
    {
        $errors = $this->inlineValidate($request, ['rules' => ['email' => 'required|email']]);
        if (count($errors) > 0) {
            return $this->response(['type' => 'validation', 'errors' => $errors]);
        }
        $user = AdminUser::where([['email', '=', $request->email]])->first();
        if (empty($user)) {
            return $this->response(['type' => 'validation', 'errors' => ['email' => [pxLang($this->lang,'mgs.no_user')]]]);
        }
        $code = mt_rand(0000000, 1111111);
        try {
            $user->reset_code = $code;
            $user->sent_at = Carbon::now()->format('Y-m-d H:i:s');
            $user->save();
            $lang = $this->lang;
            Mail::send('admin.pages.reset.mail.send-reset-code', ['code'=>$code,'lang' => $lang], function ($message) use ($request, $user) {
                $message->subject(pxLang($this->lang,'fields.email_title'));
                $message->from(config('i.service_email'), config('i.service_name'));
                $message->to([$user->email], pxLang($this->lang,'fields.email_title'));
            });
            $data['extraData'] = [
                "inflate" => pxLang($this->lang,'mgs.code_success'),
            ];
            $data['next_step'] = 'verify-code';
        $data['user_uuid'] = $user->uuid;
            return $this->response(['type' => "success", 'data' => $data]);
        } catch (\Exception $e) {
            return $this->response(['type' => 'noUpdate', 'title' => pxLang($this->lang,'','common.went_wrong'), 'content' => $e->getMessage()]);
        }
    }

     public function verifyCode(Request $request): JsonResponse
    {
        $errors = $this->inlineValidate($request, ['rules' => ['code' => 'required|numeric']]);
        if (count($errors) > 0) {
            return $this->response(['type' => 'validation', 'errors' => $errors]);
        }
        $user = AdminUser::where([['uuid', '=', $request->user_uuid]])->first();
        if (empty($user)) {
            return $this->response(['type' => 'validation', 'errors' => ['code' => [pxLang($this->lang,'mgs.no_user')]]]);
        }
        if ($request->code != $user->reset_code) {
            return $this->response(['type' => 'noUpdate', 'title' => pxLang($this->lang,'mgs.invalid_code')]);
        }
        $user->reset_code = null;
        $user->sent_at = null;
        $user->save();
        $data['extraData'] = [
            "inflate" => pxLang($this->lang,'mgs.success'),
        ];
        $lang = $this->lang;
        $data['next_step'] = 'change-pass';
        $data['user_uuid'] = $user->uuid;
        return $this->response(['type' => "success", 'data' => $data]);
    }

    public function changePass(Request $request): JsonResponse
    {
        $errors = $this->inlineValidate($request, [
            'rules' => [
                'password' => 'required|min:8',
                'confirm_password' => 'required|min:8|same:password'
            ]
        ]);
        if (count($errors) > 0) {
            return $this->response(['type' => 'validation', 'errors' => $errors]);
        }
        $user = AdminUser::where([['uuid', '=', $request->user_uuid]])->first();
        if (empty($user)) {
            return $this->response(['type' => 'validation', 'errors' => ['code' => [pxLang($this->lang,'mgs.no_user')]]]);
        }
        $user->password = Hash::make($request->confirm_password);
        $user->save();
        $data['extraData'] = [
            "inflate" => pxLang($this->lang,'mgs.pass_success'),
            "redirect" => "admin/login?pass_changed=true"
        ];
        return $this->response(['type' => "success", 'data' => $data]);
    }

    //vpx_attach

}